requirements.txt
opencv-python
pillow
flask
numpy
torch
torchvision